+++ src/manifest.ts
@@ -16,29 +16,31 @@ export default defineManifest({
     default_popup: 'popup.html',
     default_icon: 'img/logo-48.png',
   },
-  options_page: 'options.html',
-  devtools_page: 'devtools.html',
+  // options_page: 'options.html',
+  // devtools_page: 'devtools.html',
   background: {
     service_worker: 'src/background/index.ts',
     type: 'module',
   },
   content_scripts: [
     {
-      matches: ['http://*/*', 'https://*/*'],
-      js: ['src/contentScript/index.ts'],
-    },
+      matches: [
+        "*://*.rssblue.com/*",
+        "*://podcasters.spotify.com/*",
+        "*://*.buzzsprout.com/*",
+        "*://publish.blubrry.com/*"
+      ],
+      js: ["src/contentScript/hcindex.ts"]
+    }
   ],
-  side_panel: {
-    default_path: 'sidepanel.html',
-  },
+  // side_panel: {
+  //   default_path: 'sidepanel.html',
+  // },
   web_accessible_resources: [
     {
       resources: ['img/logo-16.png', 'img/logo-34.png', 'img/logo-48.png', 'img/logo-128.png'],
       matches: [],
     },
   ],
-  permissions: ['sidePanel', 'storage'],
-  chrome_url_overrides: {
-    newtab: 'newtab.html',
-  },
+  permissions: ['storage', 'activeTab'], // 'sidePanel'
 })